==> Code/4-specify-paths.R <==
# specify paths to feather files obtained in Python
path_to_scanpy_normalise_file <- "scanpy_normalise_matrix.feather"
path_to_stlearn_normalise_file <- "stlearn_normalise_matrix.feather"

# specify paths to files containing barcodes divided into spot clusters
path_to_cluster1 <- "cluster1.csv"
path_to_cluster2 <- "cluster2.csv"
path_to_cluster3 <- "cluster3.csv"

# specify path to output folder
path_to_output_folder <- "Output\\"
